pyproject.toml
[tool.poetry]
name = "disclosure-crowler"
version = "0.1.0"
description = ""
authors = ["skip"]

[tool.poetry.dependencies]
python = "^3.8"
requests = "^2.24.0"
black = "^20.8b1"
click = "^7.1.2"
loguru = "^0.5.3"
python-dateutil = "^2.8.1"
pytest = "^6.1.1"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
